Move renderer to own folder and remove RawComponent
