+++ frontend/components/elements/TextBlockComponent.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 // tslint:disable:react-no-dangerous-html
 
-export const RawComponent: React.SFC<{ html: string }> = ({ html }) => (
+export const TextBlockComponent: React.SFC<{ html: string }> = ({ html }) => (
     <span
         dangerouslySetInnerHTML={{
             __html: html,
