+++ frontend/components/lib/ArticleRenderer.tsx
@@ -1,4 +1,4 @@
-import { RawComponent } from '../elements/RawComponent';
+import { TextBlockComponent } from '../elements/TextBlockComponent';
 import { ImageBlockComponent } from '../elements/ImageBlockComponent';
 import React from 'react';
 
@@ -9,7 +9,7 @@ export const ArticleRenderer: React.SFC<{ elements: CAPIElement[] }> = ({
         .map((element, i) => {
             switch (element._type) {
                 case 'model.liveblog.TextBlockElement':
-                    return <RawComponent key={i} html={element.html} />;
+                    return <TextBlockComponent key={i} html={element.html} />;
                 case 'model.liveblog.ImageBlockElement':
                     return <ImageBlockComponent key={i} element={element} />;
                 default:
